extrapolate file extracted from graph model and put it in separate file
