+++ stable_gnn/pipelines/graph_classification_pipeline.py
@@ -10,7 +10,7 @@ from torch.nn import Module
 from torch.optim import Optimizer
 from torch_geometric.loader import DataLoader
 from torch_geometric.typing import Tensor
-
+from stable_gnn.extrapolate import Extrapolate
 from stable_gnn.graph import Graph
 from stable_gnn.model_gc import ModelGraphClassification as Model_GC
 from stable_gnn.pipelines.abstract_pipeline import TrainModel
@@ -145,6 +145,7 @@ class TrainModelGC(TrainModel):
         model.to(self.device)
 
         if self.extrapolate_flag:
+            Extrapolation = Extrapolate(model=model, dataset=self.data)
             init_edges = False
             remove_init_edges = False
             white_list = False
@@ -153,7 +154,7 @@ class TrainModelGC(TrainModel):
                 self.train_dataset,
                 self.test_dataset,
                 self.val_dataset,
-            ) = model.extrapolate(
+            ) = Extrapolation(
                 self.train_indices,
                 self.val_indices,
                 init_edges,
@@ -218,6 +219,7 @@ class TrainModelOptunaGC(TrainModelGC):
         )
 
         if self.extrapolate_flag:
+            Extrapolation = Extrapolate(model=model, dataset=self.data)
             init_edges = False
             remove_init_edges = False
             white_list = False
@@ -226,7 +228,7 @@ class TrainModelOptunaGC(TrainModelGC):
                 self.train_dataset,
                 self.test_dataset,
                 self.val_dataset,
-            ) = model.extrapolate(
+            ) = Extrapolation(
                 self.train_indices,
                 self.val_indices,
                 init_edges,
